fix: using Int64 for account_action_sequence field

+++ src/api/v1/types.ts
@@ -452,7 +452,7 @@ export class GetTableByScopeResponse extends Struct {
 @Struct.type('ordered_action_result')
 export class OrderedActionsResult extends Struct {
     @Struct.field(UInt64) global_action_sequence!: UInt64
-    @Struct.field(Int32) account_action_sequence!: Int32
+    @Struct.field(Int64) account_action_sequence!: Int64
     @Struct.field(UInt32) block_num!: UInt32
     @Struct.field(BlockTimestamp) block_time!: BlockTimestamp
     @Struct.field('any') action_trace?: any
